fix critical bug

+++ api/activetigger/languagemodels.py
@@ -9,6 +9,7 @@ import pandas as pd
 from pandas import DataFrame
 
 import activetigger.functions as functions
+from activetigger.config import config
 from activetigger.datamodels import (
     LMComputing,
     LMInformationsModel,
@@ -21,7 +22,6 @@ from activetigger.db.manager import DatabaseManager
 from activetigger.queue import Queue
 from activetigger.tasks.predict_bert import PredictBert
 from activetigger.tasks.train_bert import TrainBert
-from activetigger.config import config
 
 
 class LanguageModels:
@@ -134,10 +134,9 @@ class LanguageModels:
 
         # remove files associated
         try:
-            shutil.rmtree(self.path.joinpath(name))
-            os.remove(
-                f"{config.data_path}/projects/static/{self.project_slug}/{name}.tar.gz"
-            )
+            if name and name != "":
+                shutil.rmtree(self.path.joinpath(name))
+                os.remove(f"{config.data_path}/projects/static/{self.project_slug}/{name}.tar.gz")
         except Exception as e:
             raise Exception(f"Problem to delete model files : {e}")
